Time: 3 ms (59.69%), Space: 61.2 MB (10.46%) - LeetHub

+++ 2270-number-of-ways-to-split-array/2270-number-of-ways-to-split-array.java
@@ -1,7 +1,7 @@
 class Solution {
     public int waysToSplitArray(int[] nums) {
         
-        int[] prefix = new int[nums.length];
+        long[] prefix = new long[nums.length];
         prefix[0] = nums[0];
         for (int i=1; i<nums.length; i++)
         {
